+++ application/language/english/main_lang.php
@@ -1,6 +1,29 @@
 <?php 
 defined('BASEPATH') OR exit('No direct script access allowed');
 
+//Starter translate
+$lang['installation']  = 'Installation';
+$lang['t_fill_detail'] = 'Fill in the details below';
+$lang['host']          = 'Host Name';
+$lang['host_msg']      = 'Get it from your host, if <strong class="text-success">localhost</strong> does not work.';
+$lang['db_name']       = 'Databse name';
+$lang['mysql_user']    = 'Mysql Username';
+$lang['mysql_password']= 'Mysql Password';
+$lang['first_name'] = 'First name';
+$lang['last_name']      = 'Last Name';
+$lang['Email']          = 'Email';
+$lang['Company']        = 'Company';
+$lang['phone_number']   = 'Phone Number';
+$lang['username']       = 'Username';
+$lang['password']       = 'Password';
+$lang['password_confirm']   = 'Confirm passwrod';
+$lang['submit_install'] = 'Install';
+$lang['submit_create']  = 'Create';
+$lang['submit_go_to_app']   = 'Go To Application';
+$lang['step_1']        = 'Step 1: Database installation';
+$lang['step_2']        = 'Step 2: Admin user creation';
+$lang['step_3']        = 'INSTALLATION DONE SUCCESSFULLY';
+
 //Dashboard
 $lang['welcome_message'] = 'Welcome to Webase';