+++ application/language/french/main_lang.php
@@ -2,6 +2,29 @@
 defined('BASEPATH') OR exit('No direct script access allowed');
 
 
+//Starter translate
+$lang['installation']  = 'Instalation';
+$lang['t_fill_detail'] = 'Remplisser les détqils ci-dessous';
+$lang['host']          = 'Domaine';
+$lang['host_msg']      = 'Utilisew votre domaine, si <strong class="text-success">localhost</strong> ne fonctionne pas.';
+$lang['db_name']       = 'Nom base de donnée';
+$lang['mysql_user']    = 'Mysql Utilisateur';
+$lang['mysql_password']= 'Mysql ?ot de passe';
+$lang['first_name']     = 'Prénom';
+$lang['last_name']      = 'Nom';
+$lang['Email']          = 'Email';
+$lang['Company']        = 'Société';
+$lang['phone_number']   = 'Numéro Telephone';
+$lang['username']       = 'Utilisqteur';
+$lang['password']       = 'Mot de passe';
+$lang['password_confirm']   = 'Confirmer Mot de passe';
+$lang['submit_install'] = 'Installer';
+$lang['submit_create']  = 'Création';
+$lang['submit_go_to_app']   = 'Demarrér';
+$lang['step_1']        = 'Etape 1: Installation de base de donnée';
+$lang['step_2']        = 'Etape é: Création administrateur';
+$lang['step_3']        = 'INSTALLATION TERMINER AVEC SUCCESS';
+
 //Dashboard
-$lang['welcome_message']  = 'Bienvenue à Webase';
+$lang['welcome_message'] = 'Bienvenue à Webase';
 /* End of file filename.php */