+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ### version 0.1.1
 
 - Added Multilingual option
--Fix bug #1 user friendly installation
+- Fix bug #1 user friendly installation
 
 ### version 0.1